1.Changed generic entity id from long type to int 2.Changed all entity classes id from long to int 3.Added GenericDao and UserDao 4.Added JPA dependency to tables

+++ src/main/java/com/sonia/entities/GenericEntity.java
@@ -2,7 +2,7 @@ package com.sonia.entities;
 
 public interface GenericEntity {
 
-	public long getId();
+	public int getId();
 	
-	public void setId(long id);
+	public void setId(int id);
 }
